base_agent/requirements.txt
jsonlines
cryptography
datasets
tiktoken
pydantic[email]
pydantic-settings
python-dotenv
anthropic[vertex]==0.42.0
tabulate
openai
json-repair
rich
jinja2
fastapi
uvicorn[standard]
GitPython
diff-match-patch
swebench
duckduckgo-search
scipy
sympy
google-genai
googlesearch-python
pytest
pytest-asyncio
google-cloud-aiplatform
